Shields and Double Encumbrance UI Update
Added an Equipped Shields section with the ability to quickly toggle it on / off if it is equipped. The weight and armor are both added automatically for the player in the load and armor rating.

Double encumbrance now automatically removes the weapon damage roll buttons and adds them back in if not double encumbered.