Roll Modifiers & Item Values
ROLL MODIFIERS
-------------------
Can now be set on the character sheet by the player with a clickable toggle and a dropdown indicating the bonus. Once activated, this requires they untoggle this after it's use.

ITEM VALUES
---------------
All items now have a number input field marked Value. A value up to 5 digits deep can be entered with a minimum value of 0 being listed.